Rename `babel.new`
